api/requirements.txt
Django==1.10.3
psycopg2==2.6.2
dj-database-url==0.4.1
djangorestframework==3.4.1
drf-nested-routers==0.11.1
djangorestframework-jwt==1.8.0
PyJWT==1.4.2
factory-boy==2.7.0
fake-factory==0.5.7
Faker==0.7.3 
ipaddress==1.0.17 
python-dateutil==2.6.0
coverage==4.2


